+++ src/lib/server/index.ts
@@ -29,7 +29,7 @@ export interface RateLimiterStore {
 
 export interface RateLimiterPlugin {
   hash: (event: RequestEvent) => Promise<string | boolean>;
-  readonly rate: Rate;
+  get rate(): Rate;
 }
 
 ///// Store ///////////////////////////////////////////////////////////////////